.gitlab-ci.yml: Avoid certificate issues with curl
We currently get a failing CI job [1] with

ERROR: cannot verify gitlab.com's certificate, issued by 'CN=Cloudflare Inc ECC CA-3,O=Cloudflare\\, Inc.,C=US':
  Unable to locally verify the issuer's authority.
To connect to gitlab.com insecurely, use `--no-check-certificate'.

The reason for this is unclear. By using, which has access to the Windows
certificate store, we avoid that issue.

We have to use curl.exe instead of plain curl, as the latter is an alias
in powershell [2].

[1]: https://gitlab.com/tango-controls/TangoTest/-/jobs/3986865686