CMakeLists.txt: create shared library to export _create_TangoTest_class
In order to test PyTango functionality to run Cpp device servers, we need a shared library, that has
symbol _create_TangoTest_class exported. This function is used to create an instance of DeviceClass.
On Linux, shared library has to be named TangoTest.so hence we drop the 'lib' prefix in CMakeLists.txt
Also on Linux, PIC (position independent code) is needed for shared library and since we want to compile obejcts once,
we have to set it globally.